Update 8 - application fully functional; change of data type to double

+++ Projekt - Kalkulačka/src/model/Model.java
@@ -1,7 +1,7 @@
 package model;
 
 public class Model {
-	public long calculate(long number1, long number2, String operator) {
+	public double calculate(long number1, long number2, String operator) {
         switch (operator) {
             case "+":
                 return number1 + number2;
